+++ packages/cx-studio/cx_studio/ffmpeg/cx_ffmpeg.py
@@ -1,5 +1,5 @@
-
 from copy import copy
+import time
 from pyee import EventEmitter
 from cx_studio.path_expander import CmdFinder
 from pathlib import Path
@@ -14,118 +14,145 @@ import io, os
 import subprocess
 import concurrent.futures as con_futures
 import sys, signal
+from cx_studio.core import CxTime, FileSize
 
-import dataclasses
-class FFmpeg(EventEmitter, BasicFFmpeg):
-    def __init__(
-        self,
-        ffmpeg_executable: str | Path | None = None,
-        arguments: Iterable[str] | None = None,
-    ):
+
+class FFmpeg(EventEmitter):
+    def __init__(self, ffmpeg_executable: str | Path | None = None):
         super().__init__()
         self._executable: str = str(CmdFinder.which(ffmpeg_executable or "ffmpeg"))
-        self._arguments = list(arguments or [])
         self._coding_info = FFmpegCodingInfo()
-        self._is_running = False
-        self._canceled: bool = False
 
+        self._running_lock = threading.Lock()
+        self._running_cond = threading.Condition(self._running_lock)
+        self._cancel_event = threading.Event()
+        self._canceled = False
         self._process: subprocess.Popen[bytes]
 
-    def is_running(self) -> bool:
-        return self._is_running
-
-    def cancel(self):
-        if not self._is_running or not self._process:
-            return
-        sigterm = signal.SIGTERM if sys.platform != "win32" else signal.CTRL_BREAK_EVENT
-        self._canceled = True
-        self._process.send_signal(sigterm)
+    @property
+    def executable(self) -> str:
+        return self._executable
 
     @property
-    def coding_info(self):
-        return self._coding_info
+    def coding_info(self) -> FFmpegCodingInfo:
+        return copy(self._coding_info)
 
-    def _handle_stderr(self):
-        assert self._process.stderr is not None
-        line = b""
-        for line in StreamUtils.readlines_from_stream(self._process.stderr):
-            decoded = line.decode()
-            self._coding_info.update_from_status_line(decoded)
-            self.emit("coding_info_updated", copy(self._coding_info))
-            self.emit("verbose",decoded)
-            if 'frame' in decoded:
-                self.emit(
-                    "progress_updated",
-                    self._coding_info.current_time,
-                    self._coding_info.total_time,
-                )
-        return line.decode()
+    def _handle_stderr(self, line: bytes):
+        line_str = line.decode("utf-8", errors="ignore")
+        self.emit("verbose", line_str)
+
+        coding_info_dict = FFmpegCodingInfo.parse_status_line(line_str)
+
+        self._coding_info.update(**coding_info_dict)
+
+        if "current_time" in coding_info_dict or "total_time" in coding_info_dict:
+            self.emit(
+                "progress_updated",
+                self._coding_info.current_time,
+                self._coding_info.total_time,
+            )
+
+        if "current_frame" in coding_info_dict:
+            self.emit("status_updated", copy(self._coding_info))
+
+    def is_running(self) -> bool:
+        return self._running_lock.locked()
+
+    def cancel(self):
+        self._cancel_event.set()
+
+    def terminate(self):
+        sigterm = signal.SIGTERM if sys.platform != "win32" else signal.CTRL_BREAK_EVENT
+        self._process.send_signal(sigterm)
+        try:
+            self._process.wait(4)
+        except subprocess.TimeoutExpired:
+            self._process.terminate()
 
     def execute(
         self,
+        arguments: Iterable[str] | None = None,
         input_stream: bytes | IO[bytes] | None = None,
-        timeout: float | None = None,
-    ):
-        """
-        Args:
-            stream: A stream to input to the standard input. Defaults to None.
-            timeout: The maximum number of seconds to wait before returning. Defaults to None.
+    ) -> bool:
+        args = list(arguments or [])
+        if not args:
+            return False
+        self._canceled = False
+        self._cancel_event.clear()
+        try:
+            with self._running_cond:
+                self._process = StreamUtils.create_subprocess(
+                    args,
+                    # bufsize=0,
+                    stdin=subprocess.PIPE if (input_stream is not None) else None,
+                    # stdout=subprocess.PIPE,
+                    stderr=subprocess.PIPE,
+                )
 
-        Raises:
-            FFmpegAlreadyExecuted: If FFmpeg is already executed.
-            FFmpegError: If FFmpeg process returns non-zero exit status.
-            subprocess.TimeoutExpired: If FFmpeg process does not terminate after `timeout` seconds.
-        """
-        args = list(self.iter_arguments(True))
+                self.emit("started")
 
-        if self._is_running:
-            raise FFmpegIsRunningError("FFmpeg is already running.", args)
-
-        input_stream = StreamUtils.wrap_io(input_stream)
-
-        self._process = StreamUtils.create_subprocess(
-            args,
-            bufsize=0,
-            stdin=subprocess.PIPE if (input_stream is not None) else None,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-
-        with con_futures.ThreadPoolExecutor(max_workers=4) as pool:
-            self._is_running = True
-            self.emit("started", args)
-            futures = [
-                pool.submit(
-                    StreamUtils.redirect_stream, input_stream, self._process.stdin
-                ),
-                pool.submit(StreamUtils.record_stream, self._process.stdout),
-                pool.submit(self._handle_stderr),
-                pool.submit(self._process.wait, timeout),
-            ]
-            done, pending = con_futures.wait(
-                futures, return_when=con_futures.FIRST_EXCEPTION
-            )
-            self._is_running = False
-
-            for f in done:
-                exc = f.exception()
-                if exc is not None:
-                    self._process.terminate()
-                    con_futures.wait(pending)
-                    return exc
-
-            if self._process.returncode == 0:
-                self.emit("finished")
-            elif self._canceled:
-                self.emit("canceled")
-            else:
-                raise FFmpegError.create(message=futures[2].result(), arguments=args)
-            return futures[1].result()
-        
-    def __enter__(self):
-        return self
-    
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self._process is not None:
-            self._process.terminate()
+                for line in StreamUtils.readlines_from_stream(
+                    StreamUtils.wrap_io(self._process.stderr)
+                ):
+                    self._handle_stderr(line)
+                    if self._cancel_event.is_set():
+                        self._process.terminate()
+                        self._canceled = True
+                        self._cancel_event.clear()
+                        break
+                    time.sleep(0.01)
+            # running_cond
+        finally:
             self._process.wait()
+            result = self._process.returncode == 0
+            if self._canceled:
+                self.emit("canceled")
+            elif result is False:
+                self.emit("terminated")
+            else:
+                self.emit("finished")
+            return result
+
+    def get_basic_info(self, filename: Path) -> dict:
+        with self._running_cond:
+            self._process = StreamUtils.create_subprocess(
+                self._executable,
+                "-i",
+                str(filename),
+                stderr=subprocess.PIPE,
+            )
+
+            stream = StreamUtils.wrap_io(self._process.stderr)
+            result = self._parse_basic_info_from_stream(stream)
+            self._process.wait()
+            return result
+        # running_cond
+
+    def _parse_basic_info_from_stream(self, input_stream: IO[bytes]) -> dict:
+        result = {}
+        streams = []
+        for line in StreamUtils.readlines_from_stream(input_stream):
+            line_str = line.decode("utf-8", errors="ignore")
+            input_match = re.match(r"Input #0, (.+), from '(.+)':", line_str)
+            if input_match:
+                result["format_name"] = input_match.group(1)
+                result["file_name"] = input_match.group(2)
+                continue
+
+            time_match = re.search(
+                r"Duration: (.+), start: (.+), bitrate: (\d+\.?\d*\s?\w+)/s",
+                line_str,
+            )
+            if time_match:
+                result["duration"] = CxTime.from_timestamp(time_match.group(1))
+                result["start_time"] = CxTime.from_seconds(float(time_match.group(2)))
+                result["bitrate"] = FileSize.from_string(time_match.group(3))
+                continue
+
+            streams_match = re.search(r"Stream #0:\d+\s+", line_str)
+            if streams_match:
+                streams.append(line_str.strip())
+                continue
+        if len(streams) > 0:
+            result["streams"] = streams
+        return result